Platforms: common POSIX platform upgrader class
The LinuxProductUpgraderPlatformStrategy class was a simple clone
of the Mac version, so we can consolidate these into a common
POSIXProductUpgraderPlatformStrategy class.

Moreover, since the POSIXPlatform sets its
UnderConstructionFlags.UsesCustomUpgrader flag to false,
none of this class's methods are actually called when running
the "scalar upgrade" command, we can simply change them all to
throw a NotImplementedException().
