+++ Scalar.Common/Platforms/Linux/LinuxPlatform.cs
@@ -111,13 +111,6 @@ namespace Scalar.Platform.Linux
             return TryGetEnvironmentVariableBasePath(EnvironmentVariableBaseCachePaths, out localCacheRoot, out localCacheRootError);
         }
 
-        public override ProductUpgraderPlatformStrategy CreateProductUpgraderPlatformInteractions(
-            PhysicalFileSystem fileSystem,
-            ITracer tracer)
-        {
-            return new LinuxProductUpgraderPlatformStrategy(fileSystem, tracer);
-        }
-
         public override void IsServiceInstalledAndRunning(string name, out bool installed, out bool running)
         {
             installed = false;
